requirements.txt
flask
line-bot-sdk
apscheduler
mysql-connector-python
matplotlib
python-dotenv
gtts
pillow
